Move preconfigured api objects to a different package
