+++ orca-api/src/main/java/com/netflix/spinnaker/orca/api/preconfigured/jobs/PreconfiguredJobStageProperties.java
@@ -14,37 +14,51 @@
  * limitations under the License.
  */
 
-package com.netflix.spinnaker.orca.api;
+package com.netflix.spinnaker.orca.api.preconfigured.jobs;
 
 import java.util.Arrays;
 import java.util.List;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+/** Holds configurable properties for a Stage. */
 @Data
 @NoArgsConstructor
 public abstract class PreconfiguredJobStageProperties {
 
+  /** If enabled, a pipeline can be configured to use this configuration for running a job. */
   private boolean enabled = true;
 
+  /** Label to use on the Spinnaker UI while configuring pipeline stages. */
   private String label;
 
+  /** Description to use on the spinnaker UI while configuring pipeline stages. */
   private String description;
 
+  /** Represents stage type. */
   private String type;
 
+  /** List of stage parameters that a user can set values on. */
   private List<PreconfiguredJobStageParameter> parameters;
 
+  /**
+   * Indicates pipeline execution to wait for job completion if set to true otherwise pipeline
+   * proceeds to next stage.
+   */
   private boolean waitForCompletion = true;
 
+  /** Cloud provider with which this job interacts. */
   private String cloudProvider;
 
+  /** Credentials to use while interacting with Cloud provider. */
   private String credentials;
 
+  /** Region in which the cloud resources will be launched. */
   private String region;
 
   private String propertyFile;
 
+  /** Indicates whether this job produces any artifacts. */
   private boolean producesArtifacts = false;
 
   PreconfiguredJobStageProperties(String label, String type, String cloudProvider) {
@@ -53,11 +67,22 @@ public abstract class PreconfiguredJobStageProperties {
     this.cloudProvider = cloudProvider;
   }
 
+  /**
+   * Indicates what fields can be overridden with configured values if not present in the pipeline
+   * stage context
+   *
+   * @return list of attributes that are overridable.
+   */
   public List<String> getOverridableFields() {
     return Arrays.asList(
         "cloudProvider", "credentials", "region", "propertyFile", "waitForCompletion");
   }
 
+  /**
+   * If returned false, the stage will not be available for execution.
+   *
+   * @return true if meets validation criteria otherwise false.
+   */
   public boolean isValid() {
     return this.label != null
         && !this.label.isEmpty()
