+++ orca-clouddriver/src/main/groovy/com/netflix/spinnaker/orca/clouddriver/config/JobConfigurationProperties.java
@@ -16,7 +16,7 @@
 
 package com.netflix.spinnaker.orca.clouddriver.config;
 
-import com.netflix.spinnaker.orca.api.TitusPreconfiguredJobProperties;
+import com.netflix.spinnaker.orca.api.preconfigured.jobs.TitusPreconfiguredJobProperties;
 import java.util.List;
 import lombok.Data;
 import org.springframework.boot.context.properties.ConfigurationProperties;
