+++ orca-clouddriver/src/main/groovy/com/netflix/spinnaker/orca/clouddriver/config/KubernetesPreconfiguredJobProperties.java
@@ -16,7 +16,7 @@
 
 package com.netflix.spinnaker.orca.clouddriver.config;
 
-import com.netflix.spinnaker.orca.api.PreconfiguredJobStageProperties;
+import com.netflix.spinnaker.orca.api.preconfigured.jobs.PreconfiguredJobStageProperties;
 import io.kubernetes.client.models.V1Job;
 import java.util.ArrayList;
 import java.util.Arrays;
